fix: bug fixes and improvements
