--- src/components/header/style.module.scss
@@ -1,72 +1,0 @@
-@import '~styles/main-utils';
-
-.header {
-  min-height: map-get($spacers, '5lg');
-
-  display: flex;
-  align-items: center;
-}
-
-.header__container {
-
-  // :TODO: add to structure/grid
-  @include grid-container();
-
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  > * {
-    flex: 1;
-
-    display: flex;
-    align-items: center;
-    height: map-get($spacers, '5lg');
-  }
-
-  @include media-breakpoint-up(md) {
-    max-width: map-get($grid-max-widths, 'xl');
-    padding: 0 map-get($spacers, '5lg');
-  }
-
-  @include media-breakpoint-down(md) {
-    > * {
-      flex: none;
-
-      &:empty {
-        display: none;
-      }
-    }
-  }
-
-  @include media-breakpoint-down(xs) {
-    flex-wrap: wrap;
-
-    > * {
-      order: 1;
-      flex: 1;
-    }
-
-    // navigation
-    > *:nth-child(3) {
-      order: 2;
-      flex-basis: 100%;
-      max-width: 100%;
-      height: auto;
-    }
-  }
-}
-
-.header__logo {
-  display: flex;
-  height: 18px;
-  margin-top: -4px;
-}
-
-.header__nav {}
-
-.header__action {
-  display: flex;
-  margin-left: auto;
-  margin-right: -#{map-get($spacers, 'md')};
-}
