--- src/components/portfolio/style.module.scss
@@ -1,24 +1,0 @@
-@import '~styles/main-utils';
-
-.portfolio {
-  padding-top: map-get($spacers, '7lg');
-
-  @include media-breakpoint-down(md) {
-    padding-top: map-get($spacers, '5lg');
-  }
-
-  @include media-breakpoint-down(xs) {
-    padding-top: map-get($spacers, '3lg');
-  }
-}
-
-.portfolio__container {
-  position: relative;
-  margin-left: auto;
-  margin-right: auto;
-
-  @include media-breakpoint-up(md) {
-    max-width: map-get($grid-max-widths, 'xl');
-    padding: 0 map-get($spacers, '5lg');
-  }
-}
