+++ src/layouts/footer/style.module.scss
@@ -1,0 +1,32 @@
+@import '~styles/main-utils';
+
+.footer {
+  @include section;
+
+  padding-bottom: map-get($spacers, '2lg') !important;
+}
+
+.footer__container {
+  @include container;
+
+  & + & {
+    margin-top: map-get($spacers, 'md');
+  }
+}
+
+.footer__nav {}
+
+.footer__info {
+  font-size: $font-size-sm;
+  opacity: 0.5;
+  display: flex;
+  align-items: center;
+  line-height: 1;
+
+  svg {
+    height: 1em;
+    margin-right: map-get($spacers, '2xs');
+    stroke: currentColor;
+    margin-bottom: 2px;
+  }
+}
