+++ src/layouts/header/style.module.scss
@@ -1,0 +1,65 @@
+@import '~styles/main-utils';
+
+.header {
+  min-height: map-get($spacers, '5lg');
+
+  display: flex;
+  align-items: center;
+}
+
+.header__container {
+  @include container;
+
+  display: grid;
+  grid-template-rows: 1fr;
+  grid-template-columns: 2fr 1fr 1fr;
+
+  > * {
+    display: flex;
+    align-items: center;
+    height: map-get($spacers, '5lg');
+  }
+
+  @include media-breakpoint-down(md) {
+    grid-template-columns: 1fr 1fr 1fr;
+
+    > * {
+      &:empty {
+        display: none;
+      }
+    }
+  }
+
+  @include media-breakpoint-down(xs) {
+    grid-template-rows: 2fr;
+    grid-template-columns: 1fr 1fr;
+
+    // navigation
+    > *:nth-child(2) {
+      grid-row: 2;
+      grid-column: 1 / 3;
+    }
+  }
+}
+
+.header__logo {
+  svg {
+    height: 18px;
+    margin-top: -4px;
+  }
+}
+
+.header__nav {
+  @include media-breakpoint-down(md) {
+    justify-content: center;
+  }
+
+  @include media-breakpoint-down(xs) {
+    justify-content: flex-start;
+  }
+}
+
+.header__action {
+  margin-left: auto;
+  margin-right: -#{map-get($spacers, 'md')};
+}
