+++ src/layouts/hero/style.module.scss
@@ -1,26 +1,11 @@
 @import '~styles/main-utils';
 
-.hero {}
+.hero {
+  @include section;
+}
 
 .hero__container {
-
-  // :TODO: add to structure/grid
-  @include grid-container();
-
-  margin-top: map-get($spacers, '7lg');
-
-  @include media-breakpoint-up(md) {
-    max-width: map-get($grid-max-widths, 'xl');
-    padding: 0 map-get($spacers, '5lg');
-  }
-
-  @include media-breakpoint-down(md) {
-    margin-top: map-get($spacers, '5lg');
-  }
-
-  @include media-breakpoint-down(xs) {
-    margin-top: map-get($spacers, '4lg');
-  }
+  @include container;
 }
 
 .hero__caption {
