+++ src/pages/home/index.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { Header } from '@layouts/header';
+import { Footer } from '@layouts/footer';
+import { Hero } from '@layouts/hero';
+import { Portfolio } from '@layouts/portfolio';
+
+export const Home = () => {
+  return (
+    <>
+      <Header />
+      <Hero />
+      <Portfolio />
+      <Footer />
+    </>
+  );
+};
