+++ src/root.tsx
@@ -1,14 +1,13 @@
 import React, { useEffect } from 'react';
 import { hot } from 'react-hot-loader/root';
 
-import { useLoading } from './hooks/loading.hook';
+import { loadFontGroup } from '@utils/load-fonts.util';
 
-import { SplashScreen } from './components/splash-screen';
-import { Hero } from '@components/hero';
-import { Portfolio } from '@components/portfolio';
-import { Footer } from '@components/footer';
+import { useLoading } from '@hooks/loading.hook';
 
-import { loadFontGroup } from './utils/load-fonts.util';
+import { SplashScreen } from '@components/splash-screen';
+
+import { Home } from '@pages/home';
 
 import styleJson from './styles/main.scss.json';
 
@@ -23,9 +22,7 @@ const Root = () => {
 
   return (
     <SplashScreen isLoading={isLoading}>
-      <Hero />
-      <Portfolio />
-      <Footer />
+      <Home />
     </SplashScreen>
   );
 };
