+++ src/styles/utils/helpers/_gradient-over-text.scss
@@ -1,0 +1,29 @@
+@mixin gradient-overlay {
+  &:before {
+    content: '';
+
+    position: absolute;
+    left: 0;
+    top: 0;
+
+    background: linear-gradient(-90deg, color-rgba('background', 0) 0%, color-rgba('background', 1) 100%);
+    width: map-get($spacers, 'md');
+    height: 100%;
+
+    z-index: 9999;
+  }
+
+  &:after {
+    content: '';
+
+    position: absolute;
+    right: 0;
+    top: 0;
+
+    background: linear-gradient(90deg, color-rgba('background', 0) 0%, color-rgba('background', 1) 100%);
+    width: map-get($spacers, '3lg');
+    height: 100%;
+
+    z-index: 9999;
+  }
+}
